wsi-viewer/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
openslide-bin>=4.0.0        # bundles native openslide C library (no brew needed)
openslide-python>=1.3.1
Pillow>=10.0.0
werkzeug>=3.0.0
